Rebooting the project and starting with a fresh and very simple template
Starting out by leveraging "user-data" with a SoftwareConfig resource.
This will bootstrap the only flavor at the moment, Ubuntu 16, with apt-get update then upgrade, install docker prerequisites, then use docker to pull down the Grafana image.
Docker sets up on all interfaces port 3000. Adding a load balancer listening on 80, and pointing to the nodes 3000 port.
Currently HTTP only. Want to add HTTPS, most likely with letsencrypt or a docker plugin if available, add the original cloud metrics source, if available, fine tune the docker configuration; persistent volumes, information on accessing the container for those not use to docker.
A note on HOT syntax, I wasn't able to get "conditions" working which I wanted to use for the ability to add an existing nova keypair or create a new one. Need to look into this more but may be a limitation with Rackspace Orchestration? Not sure. More to come.